Kidney Project IA/requirements.txt
tensorflow==2.12.0
numpy>=1.20.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
pillow>=8.0.0
seaborn>=0.11.0
pandas>=1.3.0